+++ src/debug.h
@@ -1,6 +1,7 @@
 #pragma once
 
+#include "common.h"
 #include "chunk.h"
 
 void disassemble_chunk(Chunk* chunk, const char* name);
-int disassemble_instruction(Chunk* chunk, int offset);
+int disassemble_instruction(Chunk* chunk, size_t offset);
